ADDED SORTING, PAGINATION & FILTERSING FUNCTIONALITY 	modified:   src/models/task.js 	modified:   src/models/user.js 	modified:   src/routers/task.js 	modified:   src/routers/user.js

+++ src/routers/task.js
@@ -1,11 +1,15 @@
 import express from "express";
 import Task from "../models/task.js";
+import auth from "../middleware/auth.js";
 
 const router = new express.Router();
 
 //create tasks
-router.post('/tasks', async (req, res) => {
-    const task = new Task(req.body);
+router.post('/tasks', auth, async (req, res) => {
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    })
     try{
         await task.save();
         res.status(201).send(task);
@@ -15,20 +19,42 @@ router.post('/tasks', async (req, res) => {
 })
 
 //read tasks
-router.get('/tasks', async (req, res) => {
+router.get('/tasks', auth, async (req, res) => {
+    const match = {}
+    const sort = {}
+
+    if(req.query.completed){
+        match.completed = req.query.completed === 'true'
+    }
+
+    if(req.query.sortBy){
+        const parts = req.query.sortBy.split(':');
+        sort[parts[0]] = parts[1] === 'desc' ? -1 :1;   
+    }
+
     try{
-        const tasks = await Task.find({});
-        res.send(tasks);
+        //const tasks = await Task.find({owner: req.user._id});
+        await req.user.populate({
+            path: 'tasks',
+            match,
+            options: {
+                limit: parseInt(req.query.limit),
+                skip: parseInt(req.query.skip),
+                sort
+        }
+        });
+        res.send(req.user.tasks);
     }catch(e){
+        console.log(e);
         res.status(500).send(e);
     }
 })
 
 //read Task by id
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id;
     try{
-        const task = await Task.findById(_id);
+        const task = await Task.findOne({_id, owner: req.user._id});
         if(!task){
             return res.status(404).send();
         }
@@ -38,8 +64,8 @@ router.get('/tasks/:id', async (req, res) => {
     }
 })
 
-//update tasks
-router.patch('/tasks/:id', async (req, res) => {
+//update tasks with authenticated user
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body);
     const allowed = ['description', 'completed'];
     const isValid = updates.every((update) => {
@@ -53,32 +79,35 @@ router.patch('/tasks/:id', async (req, res) => {
     const _id = req.params.id;
     try {
         
-        const task = await Task.findById(_id);
+        const task = await Task.findOne({_id, owner: req.user._id});
+        if (!task) {
+            return res.status(404).send();
+        }
         updates.forEach((update) => {
             task[update] = req.body[update];
         })
         
+        await task.save();
         //const task = await Task.findByIdAndUpdate(_id, req.body, { new: true, runValidators: true });
-
-        if (!task) {
-            return res.status(404).send();
-        }
+        
         res.send(task);
     } catch (e) {
+        console.log(e);
         res.status(500).send(e);
     }
 });
 
 //delete Task
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id;
     try{
-        const task = await Task.findByIdAndDelete(_id);
+        const task = await Task.findByIdAndDelete({_id, owner: req.user._id});
         if(!task){
             return res.status(404).send();
         }
         res.send(task);
     }catch(e) {
+        console.log(e);
         res.status(500).send(e)
     }
 })
